Gives better overview of test pages by centralizing them
